+++ app/Http/Controllers/RateMasterController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\RateMaster;
+use Illuminate\Support\Facades\Gate;
+
 class RateMasterController extends Controller
 {
     /**
@@ -11,6 +13,9 @@ class RateMasterController extends Controller
      */
     public function index()
     {
+        if(!Gate::allows('view rate master')) {
+            abort(403);
+        }
         $rateMasters = RateMaster::latest()->get();
 
         return view('admin.rate-master.index', compact('rateMasters'));
@@ -21,6 +26,9 @@ class RateMasterController extends Controller
      */
     public function create()
     {
+        if(!Gate::allows('create rate master')) {
+            abort(403);
+        }
         return view('admin.rate-master.create');
     }
 
@@ -29,14 +37,24 @@ class RateMasterController extends Controller
      */
     public function store(Request $request)
     {
+        if(!Gate::allows('create rate master')) {
+            abort(403);
+        }
         $request->validate([
-            'type'    => 'required',
-            'rate'     => 'required',
+            'guard_type'        => 'required|unique:rate_masters,guard_type',
+            'regular_rate'      => 'required',
+            'laundry_allowance' => 'required',
         ]);
 
         $rateMaster = RateMaster::create([
-            'type' => $request->type,
-            'rate'  => $request->rate,
+            'guard_type'        => $request->guard_type,
+            'regular_rate'      => $request->regular_rate,
+            'laundry_allowance' => $request->laundry_allowance,
+            'canine_premium'    => $request->canine_premium,
+            'fire_arm_premium'  => $request->fire_arm_premium,
+            'gross_hourly_rate' => $request->gross_hourly_rate,
+            'overtime_rate'     => $request->overtime_rate,
+            'holiday_rate'      => $request->holiday_rate
         ]);
 
         return redirect()->route('rate-master.index')->with('success', 'Rate Master created successfully.');
@@ -55,6 +73,9 @@ class RateMasterController extends Controller
      */
     public function edit($id)
     {
+        if(!Gate::allows('edit rate master')) {
+            abort(403);
+        }
         $rateMaster = RateMaster::where('id', $id)->first();
 
         return view('admin.rate-master.edit', compact('rateMaster'));
@@ -65,14 +86,24 @@ class RateMasterController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(!Gate::allows('edit rate master')) {
+            abort(403);
+        }
         $request->validate([
-            'type'    => 'required',
-            'rate'     => 'required',
+            'guard_type'        => 'required|unique:rate_masters,guard_type,' . $id,
+            'regular_rate'      => 'required',
+            'laundry_allowance' => 'required',
         ]);
 
         $rateMaster = RateMaster::where('id', $id)->update([
-            'type' => $request->type,
-            'rate'  => $request->rate
+            'guard_type'        => $request->guard_type,
+            'regular_rate'      => $request->regular_rate,
+            'laundry_allowance' => $request->laundry_allowance,
+            'canine_premium'    => $request->canine_premium,
+            'fire_arm_premium'  => $request->fire_arm_premium,
+            'gross_hourly_rate' => $request->gross_hourly_rate,
+            'overtime_rate'     => $request->overtime_rate,
+            'holiday_rate'      => $request->holiday_rate
         ]);
 
         return redirect()->route('rate-master.index')->with('success', 'Rate Master updated successfully.');
@@ -83,6 +114,9 @@ class RateMasterController extends Controller
      */
     public function destroy(string $id)
     {
+        if(!Gate::allows('delete rate master')) {
+            abort(403);
+        }
         $rateMaster = RateMaster::where('id', $id)->delete();
 
         return response()->json([
